Move clientserver code into a subproject
Left behind was `ConnectionFactory` as it is really a MapTool-specific tool for building connections and servers rather
than a part of the general client-server logic itself. Also pulled in was `Handshake` and `HandshakeObserver` since they
are used in the client-server and are an important part of the process (the implementations were left behind though).

Common build configurations wered moved into a `shared.gradle` file applied to the root project and
subprojects. Dependencies specific to the clientserver subproject were moved out of the root project to keep things
clear.
