+++ clientserver/build.gradle
@@ -1,0 +1,33 @@
+plugins {
+    id "base"
+    id "java-library"
+}
+
+apply from: rootProject.file('buildSrc/shared.gradle')
+
+// In this section you declare the dependencies for your production and test code
+dependencies {
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.22.1'
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.22.1'
+    implementation group: 'org.apache.logging.log4j', name: 'log4j-1.2-api', version: '2.20.0'	// Bridges v1 to v2 for other code in other libs
+    implementation group: 'org.slf4j', name: 'slf4j-simple', version: '2.0.12'
+    implementation group: 'commons-logging', name: 'commons-logging', version: '1.3.0'
+
+    // Better JSON functions...
+    implementation group: 'com.google.code.gson', name: 'gson', version: '2.10.1'		// https://mvnrepository.com/artifact/com.google.code.gson/gson
+
+    // webrtc
+    implementation group: 'org.java-websocket', name: 'Java-WebSocket', version: '1.5.6'
+    // Needs to be API since WebRTCConnection implements PeerConnectionObserver and RTCDataChannelObserver.
+    implementation 'dev.onvoid.webrtc:webrtc-java:0.8.0'
+    if (osdetector.os.is('windows'))
+        implementation 'dev.onvoid.webrtc:webrtc-java:0.8.0:windows-x86_64'
+    else if (osdetector.os.is('osx'))
+        implementation 'dev.onvoid.webrtc:webrtc-java:0.8.0:macos-x86_64'
+    else if (osdetector.os.is('linux'))
+        implementation 'dev.onvoid.webrtc:webrtc-java:0.8.0:linux-x86_64'
+
+    // compression of messages between client and server
+    implementation 'org.apache.commons:commons-compress:1.26.0'
+    implementation 'com.github.luben:zstd-jni:1.5.5-11'
+}
