+++ clientserver/src/main/java/net/rptools/clientserver/simple/DisconnectHandler.java
