+++ clientserver/src/main/java/net/rptools/clientserver/simple/connection/Connection.java
