+++ clientserver/src/main/java/net/rptools/clientserver/simple/connection/SocketConnection.java
