+++ clientserver/src/main/java/net/rptools/clientserver/simple/server/ServerObserver.java
