+++ settings.gradle
@@ -16,3 +16,4 @@ include 'services:webservice'
 */
 
 rootProject.name = 'MapTool'
+include ':clientserver'
